requirements.txt
sqlalchemy
scapy-python3
pyOpenSSL
psycopg2